Fixes placement of files
